refactor(dap): extract security validation into microcrate
