+++ crates/perl-dap-security/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "perl-dap-security"
+version = "0.10.0"
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+description = "Security validation utilities for perl-dap"
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation = "https://docs.rs/perl-dap-security"
+readme = "README.md"
+keywords = ["perl", "dap", "security", "validation", "debugger"]
+categories = ["development-tools::debugging", "security"]
+include = [
+  "src/**",
+  "Cargo.toml",
+  "README.md",
+  "LICENSE*"
+]
+
+[lib]
+name = "perl_dap_security"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow = "1.0.102"
+thiserror = "2.0.18"
+perl-path-security = { workspace = true }
+
+[dev-dependencies]
+tempfile = "3.24.0"
+perl-tdd-support = { workspace = true }
+
+[lints]
+workspace = true
