+++ crates/perl-dap-security/README.md
@@ -1,0 +1,9 @@
+# perl-dap-security
+
+Security validation utilities for the `perl-dap` ecosystem.
+
+This crate centralizes DAP-focused security checks including:
+
+- workspace path boundary validation
+- safe expression/condition validation
+- timeout clamping and limits
