+++ crates/perl-dap/Cargo.toml
@@ -58,7 +58,7 @@ perl-dap-platform = { workspace = true }
 perl-dap-variables = { workspace = true }
 perl-dap-stack = { workspace = true }
 perl-module-path = { workspace = true }
-perl-path-security = { workspace = true }
+perl-dap-security = { workspace = true }
 perl-content-length-framing = { workspace = true }
 perl-keywords = { workspace = true }
 perl-lsp-launcher = { workspace = true }
